Split RecalcEngine into a base class that can live in FxCore and just… (#266)
 Split RecalcEngine into a base class that can live in FxCore and just has binding support. Tooling can use this to do binding, check results, and Move DisplayNames.

Able to directly create a SimpleResolver from a PowerFxConfig. 

Stronger type safety - a bunch of places that took FormulaType for Parameters really need a RecordType.

* Make PowerFxConfig object the source of truth.
SimplerResolver wraps a PowerFxConfig object. The resolve caches the function list, but config is the source of truth for all functions.

This is tricky because of various defaults...

- RecalcEngine - defaults to functions that interpreter actually implements
- Engine - defaults to all known builtin functions in FxCore (BuiltinFunctionsCore.BuiltinFunctionsLibrary).
- various "empty enum" tests - want an empty enum store.  But adding functions will add enums; so the only way to get a truly empty store is to also avoid adding any functions.
But some intellisense tests actually want empty enums with some functions, so had to split it up.

Solved this by allowing setting "Core functions" on the Config - these specify which default to use.