Feat(UDF): added UDF's
UDF's exist, you add them using RecalcEngine#DefineFunction and then call RecalcEngine#BindAllUDFs after adding all the UDF's. This commit also adds a limit for the number of nested function calls. This is added as a PowerFxConfig option. This commit does not include the necessary parser/interpreter changes which provides the actual syntax for defining functions.