Fixing symbol table (#503)
Symbol Table improvements:

Allow sharing config and chaining. A common config may hold all the connectors, and then a smaller config may chain to it and add a function.
Allow mutating config to enable "sessions". Mutations during bind will be caught and the binding will fail - similar to how mutating a collection during enumeration will be caught and fail the enumeration.
Allow passing in config per-expression (Rather than just engine wide)
Allow passing in runtime state per expression to functions (useful for User(), Rand(), Now()/TimeZoneInfo(), etc).
Unify globals and locals

>> Introducing SymbolTable and SymbolValues
Introduce SymbolTable as a public facing implementation of INameResolver. (This follows the pattern of FormulaType/Dtype).
SymbolTable also has a ReadOnlySymbolTable that can be handed to binding.
SymbolTable can also have a parent pointer to another SymbolTable for chaining/reuse.
Introduce a SymbolValues, which is the runtime (interpreter-only) counterpart to SymbolValues.
The SymbolValues can also provide a SymbolTable to help keep the runtime and design time in sync.
SymbolValues/SymbolTable are a generalized pattern of the previous RecordValue/RecordType used by Check/Eval.
Add Check/Eval overloads that accept these new Symbol* objects.
Much of the state on PowerFxConfig now moves to SymbolTable. To keep compact, the PowerFxConfig holds a default symbol table and methods forward to it.
**>> Simplifying CoreFunctions **
Before, we had an awkward dance where Engine would take in a config and SetCoreFunctions() to mutate the config.  This was how the engine described what functions it could support.
Now, since symbol tables can compose, the engine has a SymbolTable for whatever core functions it supports.
And that table is composed with other tables containing global state, etc.
**>>Unifying Globals and Locals **
Before, locals were passed as RowScope. Globals were handled via the INameResolver.  And there was nothing in between.
Now, everything is just "symbols", provided by a symbol table.  This did flush out some differences between rowscope and global handling.
>> This avoids breaking changes!  But does add [Obsolete] to some existing interfaces.
Notably, CheckResult.IExpression is now Obsolete - this was problematic since CheckResult is in Core, but IExpression is for evaluation (in the interpreter layer). So it was a layering violation.
Now, interpreter has an extension method on CheckResult to create a new IExpressionEvaluator, which can be used for evaluation - but both the extension method and new interface live in Interpreter where they belong.