Refactor RenameDriver to use derived record type (#591)
fixes #590 

In adding support for #590, this also refactors RenameDriver to be significantly lazier. It no longer needs to traverse the entire parameters input, just the portion actually accessed in the expression. This is done by leveraging the new lazy type support to wrap the parameter record and lazily replace the display name providers on demand.