Make PostVisitBinaryOpNodeAddition more functional (#610)
A big one, this changes PostVisitBinaryOpNodeAddition to be a mapping of inputs to outputs. This change also makes BinderCheckTypeResult a class instead of a readonly struct. This allows the Coercions property to default to an empty list instead of defaulting to null, removing the need to check for null when consuming the coercions list.