Updated single table wrapper to use the wrapped scalar function (#624)
Currently some of the tabular versions of scalar functions are not calling the scalar version properly - instead of going to the function as exposed to the user (the wrapper that handles errors and blank values), it is going to the inner implementation. This PR updates the tabular versions for single argument functions (multi-argument functions will come later).

This also fixes the Char function to treat Blank values as 0, like Excel does.