Initial work on UntypedObject ForAll (#669)
Adds the ForAll function for UntypedObjects. A very bizarre change with
some hacky type checking.

The problem is that overloaded functions with lambdas are not evaluated
bottom-up. This is because some of the arguments may be lambdas. It
leads to a chicken and egg problem: you need to know the override before
you can determine that an argument is a lambda, but you need to know
which arguments are lambdas in order to pick an override.

Thus we enforce that only a single override is allowed to have a lambda.
It is assumed to be a table, and we type check the first argument quite
late.

This implementation piggybacks off of all existing code paths by
"tricking" the type checker into thinking that there is only 1 override.
It will then "switch" the selected override at the last second if it's
discovered that the first argument is an untyped object.

Not ideal for a code maintenance standpoint. However, our other option
is to rename ForAll function to something else specifically for untyped
objects.