Add method for comparing the outputs of CheckInvocation vs CheckTypes/Semantics (#785)
The long-term goal is to split CheckInvocation into CheckTypes and
CheckSemantics, so that type reconstruction can happen at the
sub-expression level, independently of the binder. Doing this split
all-at-once is unsafe and leads to test failures in PAClient. This
change adds an opt-in mechanism, so that functions can slowly migrate to
the new methods.