Add record and replay feature to LimitedSeverityErrorContainer (#798)
Adds a "record and replay" feature to the error container that is being
used by TryGetBestOverload. This is so that, when functions want to use
CheckTypes/CheckSemantics instead of CheckInvocation, semantic errors
that do not contribute to a type checking error can be "replayed" in the
case where type checking succeeds.

A new error container is created per-function, so only the errors
generated for a single function are added to the list.

This should greatly reduce the number of failures in PAClient test labs
as we migrate functions to use the new methods.