Serializer to create human-friendly representation for .txt test cases (#782)
Currently there are 2 types of "serialization": the official one (if I
can say that) implemented a few weeks ago use and the one that it is
used to translate a test case FormulaValue result into a human readable
value. The idea is to unify both result formats into a single one,
provided by the official serializer.

Looking for suggestions/comments. Things that are still under
consideration:
- How to represent ErrorValues? Currently there are mainly 4 formats
depending on the case:
  - Microsoft.PowerFx.Types.ErrorValue
  - #Error
  - #Error(Kind=InvalidArgument) or #Error(Kind=InvalidArgument, Div0)
- Errors: Error 0-11: The function 'Boolean' has some invalid
arguments.|Error 8-10: Invalid schema, expected a one-column table.
(This format most probably will remain since it display useful
information)
- Single column table type is now serialized and displayed as
**Table({Value:1}, {Value:2}, ..., {Value:N})**
- Empty single column table is now serialized and displayed as
**Table()**

There are still a few errors to cover.