Update static format check for Text function (#866)
The Text function has a static check that the format argument should not have both numeric and date/time specifiers. This logic will change (once we support `0` as the specifier for fractional seconds, as Excel does), and it isn't implemented correctly. This change updates it to cover more constant values. Notice that this should not be a replacement for a runtime check, since the format can come from expressions, variables, which are not captured by a static check.