Only treat an inlinerecord field as a member of a scope if it's not marked SkipForInlineRecords (#881)
`ForAll` and other functions can be marked as SkipForInlineRecords.
Those functions should not have their scopes used for display names
associated with Inline Records. This is a fix for a long-standing bug in
PowerApps where we incorrectly rewrite inline records on display name
changes, even when they should not be associated with any display names.