Color primitive; ColorValue, RGBA func (#706)
Last week was the second time I ran into wanting to use Colors, and them
not being implemented. My previous PR was stalled due to changes in
marshalling and tests and I never picked it up to redo it.

So this PR replaces the old. It also adds ColorValue, ColorValue_UO and
RGBA functions.

One open question... ColorValue on PowerApps supports naming CSS colors
besides using the '#' color notation. .NET has a Color.FromName function
but unfortunately in .NET Standard does not support the IsKnownColor
property so there's no way to tell if the name was valid or not... So
right now ColorValue/UO in this PR only supports the '#' notation which
is not consistent with Power Apps.

Co-authored-by: jas-valgotar <jasvalgotar@microsoft.com>