Updated multiarg tabular function handler to return first error or blank tabular arg (#1064)
Currently the implementation of the tabular function handler needs to make two passes over the function arguments to determine whether a blank or error value will be returned (if passed in a tabular context). This is inefficient, so we are updating the implementation to return the first of those values.

This change also simplifies the implementation of the tabular handler, to avoid copying the arguments multiple times.