Add rounding to trig table tests (#1073)
The trigonometric tabular tests use functions that don't return whole numbers (such as Pi() and other trig functions). This can make the results not match exactly depending on floating point behavior, so we're adding a Round call to those tests to make sure that the results are the expected ones in all platforms.