Refactor tabular tests for PAClient runner (#1079)
As part of the excel compat work, tabular overloads of functions will
return "Value" as the output column name. This PR just refactors tabular
test coverage so that all tests with the old behavior are consolidated
under exclusive files.

- ArgCoercionT.txt
- CharT.txt
- ConcatenateT.txt
- Distinct.txt
- ExpT.txt
- FindT.txt
- LenT.txt
- LnT.txt
- RoundTable.txt
- Split.txt
- SqrtT.txt
- TabularOverloadsBlanksAndErrorsT.txt
- TableStringfuncsT.txt

The _ConsistentOneColumnResult versions cover the same test cases as
these files, but with the new behavior (Value as output column name).