Restrict IsEmpty arguments to tables only (#1236)
IsEmpty is supposed to check if a table is empty or not, but currently it accepts different types of arguments. Since it doesn't make sense to use it for other types, we will restrict it. But to avoid breaking Power Apps, we will make the change behind a new feature.

Also adding the implementation to the C# interpreter.