Update .txt tests to use the features enabled in Power Apps by default (#1228)
We currently have tests that run in Power Fx but are being ignored in Power Apps, because the JS runner doesn't understand some of the #setup directives. This updates the tests to use the same defaults as Power Apps, so that the tests won't have any setup directives, and will run properly in the Power Apps repository.