Add IR printing with type and Import improvements to Console REPL (#1231)
Adds a pseudo-function IR() to the REPL which is very handy to see the
IR for an expression. Added type information which is also very handy to
have. Note that the expression does not need to be quoted.

```
> ForAll([1,2,3], Mid("World", Value, 1))
["W", "o", "r"]

> IR( ForAll([1,2,3], Mid("World", Value, 1)) )
ForAll:*[Value:s], Scope 1(Table:*[Value:n]({Value: 1:n}, {Value: 2:n}, {Value: 3:n}), Lazy(Mid:s("World":s, ScopeAccess(Scope 1, Value), 1:n)))
```

Totally open to feedback on the new format.

All changes here are not Power Fx maker facing. This change opens up a
small amount of public surface area CheckResult.PrintIR() which returns
the IR pretty printing. I don't think this would be an issue, but wanted
to call it out.

Also adds a second argument to Import to write out the results of the
run. Can be used to feed REPL with a raw list of tests, and it will emit
the correct output for making it into a test .txt file.

ReflectionFunction was changed to support more than one function with
the same name, providing the ability to do parameter overloading, used
here for a 1 arg and 2 arg version of Import.