Enable async delegation (#1264)
This feature looks to enable async and impure node delegation as long as
Features are enabled for it.

Additionally, removes the ICustomDelegationFunction interface in favor
of overriding the delegation validation strategy in some functions to
always delegate async calls of the function for appropriate functions
(i.e. current behavior of AsType or User function in Power Apps).