AggregateCoercionNode  bug. (#1271)
We only add fields to AggregateCoercionNode that require a coercion. But
eval only evals the fields listed in AggregateCoercionNode. So a field
that doesn't require coercion gets skipped in eval!

This fixes it via eval - but should probalby be fixed at IR level. Since
eval processes every time, whereas we know this statically.

---------

Co-authored-by: Mike Stall <jmstall@microsoft.com>