Updates to Decimal support (#1293)
Added more tests, updates to testing infrastructure, minor bug fixes and
tweaks, and change to Date-Date and Time-Time to return Decimal when
NumberIsFloat is disabled. Rules for when a decimal vs number is
returned, when involving a decimal, have been simplified. Over half of
the file changes are adding tests.

I later discovered that there is a breaking change here for
NumberIsFloat. In main currently, Today()*Today() and other math
operations involving dat/time is a compile time error, a regression.
Only for NIF. That has been corrected with this PR.

Only for the interpreter. When doing Blank()-Today(), the error return
used to be `Invalid Argument - Runtime type mismatch` which is really a
missed internal check. With this change, it now properly reports
`NotSupported - The operator is invalid for these types`. We had missed
a place where we should have been replacing blank values.