Update CheckColumnType to incorporate return type for function result (#1311)
This PR fixes https://github.com/microsoft/Power-Fx/issues/1252 which
has a description of the issue.

Turns out most of functions that take a single column table had this
bug. MathFunction.cs had a good solution to the problem, which I had
initially propagated to the other functions. After feedback from the
team, I generalized the code to move this solution into the
CheckColumnType functions, significantly simplifying the CheckTypes in
each of the functions.

This fix is needed for Round* and Trunc in particular, because with
decimal single column tables these functions would hit an assert and
stop the test suite from proceeding. Something as simple as [1,2,3]
would be a problem.