Support decimal return values for integer returning functions (#1378)
A variety of our functions return simple integers, such as Len, Count,
Year, Hex2Dec, etc. At present, they all return a Number, even if
NumberIsFloat is disabled. This means that calculations involving them
will be promoted to Float and there could be a loss of precision. This
PR addresses this by returning a Decimal for these functions if
NumberIsFloat is not in effect.

Addresses https://github.com/microsoft/Power-Fx/issues/1346.

---------

Co-authored-by: Luc Genetier <69138830+LucGenetier@users.noreply.github.com>