Reversing the default for ReservedKeywords (#1408)
In general, new features shouldn't change default behavior, they should
be an opt-in. However in this case it makes sense to make it an opt-out,
controlled by an internal flag that Cnavs can set. The nature of this
feature and that it is set in ParserOptions and not directly effected by
engine.Features (as there may not be one) led us to this conclusion.

Fixes #1385