Only allows coercion to first argument type on tables (#1489)
We have update the Table function (and the `[...]` syntax) to favor the first type. But the current union logic would allow coercion from the first value into subsequent ones, which would cause issues like in the example `[ErrorKind.Div0, 12]` - the resulting type is `![Value:ErrorKindOptionSet], but the IR would try to coerce the data to number (since numeric option set can be coerced to number, but not the other way around). This change forces, within those contexts, that the union should only happen to the left type, not the right one.

Fixes #1477