Adding the very small part of logic behind hiding tokens I forgot to migrate in PR #1420 (#1510)
The support for semantic tokenization was added in PR
https://github.com/microsoft/Power-Fx/pull/1420. I migrated the logic
behind tokenization in old formula bar from Canvas App backend to
PowerFx. In doing so, i forgot to bring over a very tiny part but
important for old formula bar of the logic that computes whether first
name tokens can be hidden or not.