Make the Switch/IfError functions follow the first-type rule (#1521)
Currently the Switch and IfError functions use an intersection rule for the arguments - Switch(var, 1, {a:1}, {b:2}) returns an empty record since there are no common properties in the two records. With the PowerFxV1 rules, we are changing it to follow the "union with first-type rule" that is used in other functions, so the type of that expression would be ![a:n] with the new feature.

This is similar to #1507 which did the change for the If function.