Adding ability to supply custom binding config to Engine.ComputeBinding() & Fixing Dup String Interp Tokens bug (#1531)
This pr handles few things that are mentioned below:
1) It adds a new virtual overload of `GetDefaultBindingConfig` in
Engine.cs that takes in ParserOptions and optional rule scope and
returns a binding config that is used by `Engine.ComputeBinding()`. This
is done so that Power Apps can supply custom binding config for semantic
tokenization. Power Apps never set
`BindingConfig.UseThisRecordForRuleScope` to true but changes before
this pr calculated the value of
`BindingConfig.UseThisRecordForRuleScope` based on the value returned
from `Engine.GetRuleScope()`. It would be true when GetRuleScope()
returns a non null value. There are properties such as `OnErrors `on
AppInfo control in PowerApps that returns a non null value for
`GetRuleScope()`. In this case,` useThisRecordForRuleScope `would be
true in` Engine.ComputeBinding() `but it should never be true when
`Engine.ComputeBinding() `is run from PowerApps.

2) This pr fixes a bug where duplicate string interpolation and island
tokens were added when expression has nested string interpolation to the
list of tokens returned from Tokenization.Tokenize. This pr fixes that
logic and adds tests to validate it. A good example of this bug is an
expression like `$"{$"{$"{$"{.12e4}"}"}"}"`. The number of tokens should
be 17 but it is 41. Many of them are duplicate with same spans and token
types.

3) This pr also fixes one test that happened due to Visual Studio
picking up a different line feed for C# multiline strings `@""` .
Updates the test to avoid using multiline string. @gregli-msft is the
only and first one so far to face this problem. Tests were failing for
him. Confirmed with him that avoiding multiline string fixes the test.