Added partial collect tests (#1538)
## Issue

This is the issue tracking this:
https://github.com/microsoft/Power-Fx/issues/1543

## Change description:
This PR adds a partial collect tests using only PowerFx recalEngine. The
following is an explanation of the tests added by the this PR:
1. Powerfx_Collect_Without_Serielize
a. Defines NewTable FormulaValue for a collection with a defined
RecordType as schema type and list of NamedValues as value
b. Defines recalEngine and symbol tables and runs
`Last(collection).Col1` on it
		c. Used recalEngine to get the FormulaValue of the collection
d. Define recalEngine and symbol table and runs `Collect` with a subset
of the column keys of the collection
e. Used the same recalEngine as above to run `Last(collection).Col1` and
verifies the correct result
2. Powerfx_Collect_With_Serielize
a. Defines NewTable FormulaValue for a collection with a defined
RecordType as schema type and list of NamedValues as value
b. Defines recalEngine and symbol tables and runs
`Last(collection).Col1` on it
c. Used recalEngine to get the FormulaValue of the collection
(FormulaValue type = `RecordsOnlyTableValue`)
		d. Serielizes `RecordsOnlyTableValue` 
e. Define a recalEngine to de-serielize the `RecordsOnlyTableValue`. We
notice the output of de-serielization has the FormulaValue of type
`InMemoryTableValue`
f. Use the above recalEngine to run `Collect` with a subset of the
column keys of the collection
g. Used the same recalEngine as above to run `Last(collection).Col1`.
This leads to an error of the type `Runtime type mismatch`

---------

Co-authored-by: Mike Stall <mikestall@hotmail.com>