Filter function limited to two args (#1625)
Addressing https://github.com/microsoft/Power-Fx/issues/1514.  

Advantages:
- This brings Filter and LookUp into alignment.  
- Using And to combine predicates is easier to understand and read than
knowing that extra arguments are And'd together.
- Limiting to two arguments allows us to add a reduction argument like
LookUp has. Which enables us to push shaping of the result up the stack
more easily when making a network call.

For C# clients, the Filter function was already effectively limited to
two args because the interpreter couldn't handle it (a bug), so no
change there.

For Canvas, this is a breaking Power Fx 1.0 change (under Power
FxV1CompatibiltyRules flag) that we'll need to manage later, but only
~13% of filter calls use the three or more argument form. A back compat
converter should be possible.