Implement Abstract DelegationHintProvider (#1661)
In the upcoming months there are plans to add multiple warnings which
pertain exclusively to Canvas into Power Fx. While we decide on a
strategy for allowing Canvas to extend the warning capabilities provided
by the Binder, I am proposing the creation of a new data structure that
extends the allows consumers to add custom delegation hints by providing
an abstract class which implements virtual methods and invokes them
during delegation warning visitation on an instance provided baller the
caller on TexlBinding.Run.

Currently, the data structure implements two virtuals, one to visit
callndoes and one to visit binary op nodes. This decision was a
consequence of the needs of [this
PR](https://github.com/microsoft/Power-Fx/pull/1627) is reaching an
imminent deadline. The call node visitor needs extra information from
the binder, which is curious, and while writing this I believe we may
want to simply do a second visitation over some rules in Canvas rather
than extending the binder. However, this data structure gives us a means
of keeping the new offline-profile delegation warning implementations
out of the open source while we develop a long term plan that is most
extensible for our users.