Taking expression out of document uris and moving it to a separate property in the LSP payloads (#1655)
It was recently discovered that any expression beyond ~30K characters
essentially breaks down the language server and all the editing features
no longer work. This is due to the fact that expression is included in
the document uri which itself is wrapped into json rpc payloads which
are then sent to language server from the clients such as new formula
bar. The parser used today to parse the Uris enforces that maximum
length constraint and language server cannot do anything if that parser
generates an exception when uri is length is beyond the maximum length.
This pr attempts to move expression out of the document uris and into
its own separate property in the json rpc payloads. I did not add any
new property but instead used "Text" property that was already present
in the extended version of TextDocument. I refactored the object models
to pull common properties into base classes to avoid repeating them. I
changed the type of TextDocument to its extended version which already
had the "Text" property to represent the expression. This extended
version is used for custom initial fixup protocol. Additionally, i
noticed that an option was added to allow consumers to pass in custom
logger Action/callback. This callback is an optional and defaults to
dummy function. The log strings generated include payloads sent to
language server and all of the suggestions in the case of completion
requests. This is dangerous thing to do as raw payloads contain PII
information and adding all the suggestions in the log string is very
expensive as there are usually big number of suggestions. We would be
emitting huge strings in those cases. Temporarily, i did not default
logger callback to a dummy function as it would still result in creation
of these huge log strings with no effect. I kept logger null and only
invoke it when it is not null, thus avoid creating those hug string when
we don't need to. This way consumers of language server who do not use
this logger callback are not hit by this performance degradation by
default. But we should consider reverting these changes or coming up
with better approach.

Note that this pr only makes language server work with large expressions
like the ones with 30K characters or more. This doesn't mean language
server is faster when dealing with these large expressions. It is infact
slow and we need to come up with a better way to exchange the larger
expressions between language server and its clients.

The solution in this pr is backwards compatible and i checked with the
cards team and confirmed that changes made in this pr would not break
their scenarios

Update: I originally changed the type of "TextDocument" field on the
params object to its extended version which had more fields including
"Text" which represents an expression. I was seeing Public API Changes
warning in the pr so i changed the type of "TextDocument" field back to
its original version and added a separate property called "Text" at the
same level as "TextDocument". The object models used by Language Server
class should never have been public and should completely be internal
like the ones are for Semantic tokenization. But i didn't convert them
to internal for now as cards backend code still refer them to directly
and didn't wanna break their compilation.

---------

Co-authored-by: Luc Genetier <69138830+LucGenetier@users.noreply.github.com>